Move module tests into test scripts, update context comment
There was some test code which was not being run for the module system
which actually lived in the module boot code file.  This moves that
into the test script so it runs, and takes out the tests that don't
work.

Also updates the comments on context to reflect the current layout.