FORM non-STRING! (instead of non-ANY-STRING!) in FIND
Rebol issue 1160 pointed out a problem with using a TAG! to search
over a string:

https://github.com/rebol/rebol-issues/issues/1160

This change uses FORM on the argument, so `find "<abc>" <abc>` will
account for the tag delimiters.

Searching has some strange ramifications for other types, given that
you can't really get a good result out of `find <abc> "<"`.  For
lack of a better rule, though, this goes ahead and FORMs.

Also gets rid of the unimplemented wildcard features in FIND and
SELECT, which are more suited to a PARSE or GLOB dialect.