Eliminate Pick_Path() form of GET-PATH!/SET-PATH! reuse
Path processing (which is notoriously ad-hoc) had the mechanical code
for implementing PICK and POKE for some types, notably TUPLE!, DATE!,
IMAGE!, TIME!, and VECTOR!.  Because the code was present in the path
dispatch function `PD_Xxx()` for these types, rather than rewrite it
again for PICK and POKE they would call a routine called "Pick_Path()"
which would do a primitive form of 2-element path dispatch.

This commit reverses matters to where these types have separate Pick()
and Poke() functions, and it is those functions that are used by both
PICK and POKE as well as called from within the path dispatch.  While
this is no panacea for imposing sanity on path semantics in general,
it at least gets rid of Pick_Path().

Includes some tidying in the process--including fixing a bug in floating
point vectors and making the date code a bit clearer, by switching on
symbols instead of on integers.