fix Pick_Vector for decimal vectors
@HostileFork changed a reset header to REB_DECIMAL with an
INIT_DECIMAL_BITS() to a single call to SET_DECIMAL.  This was not
correct, because the "bits" initialization of decimals is using a
non-standard-C trick of assuming a decimal can be represented using
a 64-bit integer bit pattern via a cast.

Ideally this assumption would not be made in the code, but so long
as it is...this reverses the change as well as adding in a check in
the C++ build which makes SET_DECIMAL not accept integers directly.