fix Mold_Vector (plus more unification of series termination)
@guiliolunati noticed that ASSERT_SERIES_TERM was not satisfied when
removing the final space in Mold_Vector.

@HostileFork expanded this to have more unification of SET_SERIES_LEN
and TERM_SERIES, with specializations for certain known series
widths.  This is because generic series termination uses memset() to
zero of the tail unit, and often it is known what size series is
intended at the callsite.

(The termination behavior in Rebol in general is a fairly aggressive
micro-optimization.  Series are not kept in a terminated state
at all times but must become terminated before certain operations.
This has been an ongoing effort to make less error prone while still
attempting to keep the optimization going.)