Make ECHO return void, revert fixture line ending change
The return result from the ECHO native was given as R_OUT (R_RET in
R3-Alpha), meaning "whatever is in the output cell location".  But
it never wrote the output.  This was tolerated in R3-Alpha but is
not in Ren-C, so this changes it to return a void.

Also reverts the fixtures files which had their line endings normalized
becaues this changed the binary content, which breaks tests.