Do not setup FFI types in struct! construction
Even though the existance of struct! is for FFI, not every use case
requires a FFI type for it. e.g. passing a struct! pointer in/out of a
function.

This commit also changes the storage of REBVAL as a byte series, so it
could be any size (no worries about the memory alignment).