Error in PARSE on unset values, do not evaluate for QUOTE
At some point, using a non-keyword stopped triggering errors, instead
attempting to match a void in the input (not possible, since voids
cannot be put in blocks).  This adds an error message after the
variable fetch.

Also, QUOTE of a GROUP! was evaluating the group.  This behavior is
contrary to one's expectation of what QUOTE would do, and Red also
does not evaluate.  This removes the evaluation.

Some trailing whitespace is trimmed from comments.