Factor redundant code in Manage_Series/Free_Series
Prior to the integral use of inline functions, the Free_Series and
Manage_Series routines manually inlined a small bit of code.  This
defers to the compiler's inlining.

Also removes the compiler-flag for whether UTF-8 is being used, as
all builds will use UTF-8 (and there is no intent to support a
non-unicode build in the future).