Efficient TIGHTEN generator for making OP!-like stubs
This makes `+: enfix tighten :add` a faster version of:

   +: enfix func [arg1 [<tight> any-value!] arg2 [<tight> any-value!] [
       add :arg1 :arg2
   ]

It avoids the overhead of the user code stub, via direct reuse of the
function's implementation (just with an updated parameter list).  Also
the parameter types and help notes will match the original.

Currently this cannot be used on specialized, adapted, or chained
functions.  However, the `<tight>` feature is likely to be phased out
in Ren-C, so this is not likely to be a priority.

(The `<tight>` annotation was introduced while trying to define a
bridge for compatibility with R3-Alpha's OP!.  That code made use of
"lookahead suppression" on the right hand side of infix operators, in
order to give a left-to-right evaluation ordering in pure infix
expressions.  After some experimentation, Ren-C came up with a more
uniform rule across both "enfixed" expressions (of arbitrary arity) and
ordinary prefix expressions, which still gave a left-to-right effect
for binary infix ops.  This form of dispatch is likely to be the only
form supported in the future.)