Eliminate bad cast usage in HTTPS (breaking read on matrix.org)
The internal structure of a REBSER node was changed inside the core.
An invisible dependency on the layout existed in %host-core.c, which
was exposed when certain encryption methods were used:

https://github.com/zsx/r3/blob/e95a54cb5eb7a5fc38054547879e15eddfb27bbb/src/os/host-core.c#L531

    //hack! - will set the tail to key size
    *((REBCNT*)(((void**)priv_key.series)+1)) = dh_ctx.len;

This was attempting to set the length of the series, by assuming that
the series length could be found one 32-bit-pointer's width inside
the data structure.  This would not have ever been true on 64-bit
systems, and is obviously a very sketchy thing to be doing...if the
internals of the data structure ever changed.

This replaces the corresponding calls with the now-exported API for
RL_SET_SERIES_LEN (although the "RL" API is being retooled to speak
in terms of lifetime-managed REBVALs only, and not raw series.)  It
also fixes some memory leaks.

(Note: %host-core.c is an antiquated and scary piece of code, which is
slated for complete replacement.)