Re-relativize function bodies for interpreted functions in TIGHTEN
The starting implementation of TIGHTEN made a copy of the paramlist
of the original function without making a copy of the body.  However,
this isn't correct...because interpreted functions have bodies that
are relativized specifically against their parameter list, to know
their identity when interacting with specifiers.

Plus, even a native's singular body needs to be copied, if the copy
can have a unique behavior for HIJACK and other dispatcher-influencing
functions.  Otherwise, HIJACK of :+ would also HIJACK the :add that
it was a tightening of.  That's not an essential feature, but good
to think about as this is the first example of a function parameter
tweaking primitive in Ren-C.