Correct error for unsupported read https redirect
FIND returns a series position on success, which means a BLANK! can
be used to indicate failure...since even if you were searching for
a BLANK! then you would have gotten the position of the BLANK.

However SELECT returns a value, and BLANK! is a legitimate value.
So void (not a legitimate value) is used to signal a failure in Ren-C.

This corrects a case where SELECT was being used to determine if an
object contained a member, and returned void (as opposed to the NONE!
in R3-Alpha).  Substituting with FIND allows it to raise the correct
error.

NOTE: There is a memory leak in this error condition of memory out of
the OS allocator.  This is not detected by ordinary debug builds, but
valgrind or address sanitizer notices it.  The issue is in the
%host-core.c code whose implementation and methodology are slated
for complete replacement.