Replace %host-core.c with %n-crypt.c
This drastically cleans up the brittle RL_API-based encryption code to
make it look like fairly normal natives, in a move to pare down the
"COMMAND!" dependency.

In the process, it switches the handles used to hold the AES and RC4
cipher streaming state to being managed.  This means that freeing them
is just a matter of not holding any outstanding references.  Hence
they will not show up as malloc() leaks on various network failures.