Hardcode RXT_xxx enum to not require Rebol generator code
The RXT_xxx numbers were intended as a parallel and more stable
alternative to the REB_xxx numbers.  But rather than just hardcode them
in the extension API file, there was Rebol code that would process a
table in a somewhat obtuse way to produce them.

This gets rid of that process and just initializes tables in C code.