/ASTRAL refinement for TO-STRING
The current implementation limits codepoints in strings to 2 bytes in
size maximum (the "standard" unicode plane).  This means it's not
possible to represent codepoints in the higher planes, which are
humorously referred to as the "astral planes".

Historically these were not very commonly used, and were for more
esoteric historical languages (egyptian linear B).  The growth of
emoji usage has made them more common in Internet communication.

As a stopgap to full support, this makes TO-STRING a native which can
accept a handler.  This handler can substitute other string or
character values for the high codepoints, e.g.:

    >> test-data: #{22646973706C61796E616D65223A224A6F6E20F09F98BA22}
    == #{22646973706C61796E616D65223A224A6F6E20F09F98BA22}

    >> decode-cat: func [x] [
        if x = 128570 [return "[smiling cat with open mouth]"]
    ]

    >> to-string/astral test-data :decode-cat
    == {"displayname":"Jon [smiling cat with open mouth]"}

Literal characters or strings can also be passed as the parameter to
the /astral refinement.