Build patches on HaikuOS for gcc 2.95
This allows building on HaikuOS with the distributed GCC 2.95 that
comes with the R1 alpha build, but must be built in C++ mode.

That version of GCC (over 15 years old) does not include options for
a C build to have declarations directly after statements.  That ability
was in C++ since its inception, and was formally added to C in the C99
standard.

There were a few errors resulting from compiler bugs regarding the
type coercion of ?: ternary operators.  Since there were so few and
accomodation so minor, it was added to get the build working.

On the virtual machine tested, the compiler hit a problem when the
debug build was used.  So -DNDEBUG was also necessary.

The motivation for this is primarily to examine the dependencies of
the build, to see if it can still be built on older and more basic
platforms, using older compilers.