Some cleanups, comments, and refactoring of %sys-core.h
Historically %sys-core.h has been somewhat monolithic without a real
explanation of why it is laid out the way it is.  Ren-C had to do a
large refactoring in order to enable the use of inline functions,
which made things a little more clear but was still fairly disorganized.

This makes some minor improvements, trying to put more definitions
where they belong.  For instance, there is no reason to define the
fail() and panic() macros before %tmp-funcs.h, because no inline
function can use them until the prototypes for Fail_Core() and
Panic_Core() are available...so they can be moved afterward, and even
better into the %sys-trap.h header where their relationship to
PUSH_TRAP() can be better explained.

Also gets rid of the GC_Mark_Hook, as the new concept for external
code participating in garbage collection is to make use of "pairing"
series.  Ren-Cpp was already retrofitted to use this.