Patches for %gtk.r, %varargs.r FFI samples
A couple of minor points were causing problems with the FFI samples.
One is that all function parameter lists must now carry the
ARRAY_FLAG_PARAMLIST bit... the FFI function generators did not use
the ordinary paramlist generator so they were missing that.

Another thing was that the VARARGS! value passed to the FFI functions
did not have a managed frame value.  This could be a problem if the
FFI code happened to turn around and trigger Rebol code that caused
a GC, because it would reach an unmanaged value.

(This is worth noting also if any natives are written which take
VARARGS!...they will need to manage the frame explicitly, since there
is no dispatcher to do it for them.  They will be caught by the same
assert that caught this.  It could be avoided if one remembered that
a varargs was seen and then managed the frame in the evaluator, but
that puts the burden of cost on every function call...not just the
variadic ones.)