R_FALSE/R_TRUE cast from REBOOL, use typed REB_R enum
This adds the UNSET? native (which is then disabled for transitional
purposes), but also switches the REB_R constant values to allow for
C and C++ booleans as integers to be used directly as the REB_R return
result from natives.

The change made a bug in CHECKSUM apparent where it was returning a
0 instead of R_OUT.  So a typed enum audit was used to find other
violations, of which there were a few.  Considering how easy a mistake
it is to make, this uses the typed enum in the C++ build instead of
a fixed-size type, because the risks outweigh the benefits otherwise.