Remove unused code from %make-boot.r
In the interest of paring down as much unnecessary code as possible,
this removes the Mold and Form dispatch table generation -- which was
commented out and not used in R3-Alpha in 2012 and still not used.

Also takes some test code that was in the %mezz-save.r file and being
stripped out and turned it into active actual test code.  Then got
rid of the test-stripping in %make-boot.r, as there's not much point
to adding un-exercised test code to mezzanine files when there are
actual tests.