Make Date host API fill a date REBVAL, not arbitrary struct
There were two aspects to the "host kit".  One was a library for
calling Rebol services from C...the so-called Reb-Lib-Api, or "RL_API".
The obvious service was RL_Do_String(), which would LOAD and DO a
C string of UTF-8 text.

The other aspect was services provided by the host.  This covered
things such as getting the system time, starting processes, generating
UUIDs, etc.  Any time a platform-specific service was needed, a new
function would be added into the table that a host needed to provide.

(Note: One of the most generalized hooks was the provision of code
for creating requests for "rebol devices", and then polling or
dispatching to these devices.  This was similarly ad-hoc.)

Because Rebol was closed source, it was assumed that the code providing
the OS services did not have access to Rebol's "internal API".  This
meant that if data were to be exchanged between the interpreter and
the host, the Rebol side would have to formulate its query parameters
using something other than Rebol data.  And the result would be
handed back in the form of something other than Rebol data.  Then the
Rebol side would have to put it together into a Rebol value.

An example of this runaround was shown by the existence of a structure
called REBOL_DAT.  This was *not* how date data was encoded internally
to Rebol, rather it was a go-between C structure for passing the
various month/date/time/nanoseconds/timezone fields between the OS
get-time service (for NOW or reading times off files).

As a first step toward simplification, this commit switches to use an
actual REBVAL date to exchange this information.  A small RL_Api
routine to make a date is added; but the internal API could provide
this functionality, as could more generalized DO-based RL_Apis.

This also gets rid of a bad dependency of the scanner on the clock,
where the current year is used to interpret short dates...e.g. 96=>1996
if your clock is closer to 1996 than it is to 2096.  Instead just goes
ahead and considers it to be a short date, and it's up to clients
to fix up dates with `year < 100` if they know that's wrong.