Change Scan_XXX routines to take output as first argument
As a general coding convention, if there's a mutable output for a
function call that is the first argument.  Also, naming it "out" is
helpful if the main purpose of the function is to produce that
output.

This makes that adjustment for the Scan_XXX routines, as well as
having all of them scan REBVAL* (as opposed to having Scan_Integer
generate an INT64).  Includes some general code tidying, as the
scanner code is very old and terse.