Get rid of tabs in %make-make.r
The somewhat archaic GNU Make format requires the use of tabs.  Because
a string literal in %make-make.r was used to produce the makefile,
the whole source file inherited the legacy of using tabs...unlike the
rest of the codebase.

This switches it around so that the tabs are introduced as part of the
output process.  Any series of four spaces are written out to the
makefile as a tab.  However, the source code file that produces the
makefile can avoid having any literal tab characters in it.