Fix specific binding for PARSE rule path evaluation
During the PARSE, it maintains a specifier for the current rule array
in P_RULE_SPECIFIER.  This is the correct specifier to use with P_RULE,
but if one goes deeper into that (such as with a GROUP! that is in the
P_RULE) then it may be that the specifier is overridden if that value
itself is specific.

This also improves the error message given when a specificity mismatch
is foudn.

Includes some residual format changes to %make-make.r that were left
out of the last commit.