Working implementation of stack-frame GC'd pairings
When it was made possible for "singular" Rebol Arrays to hold a value
inside the REBSER node itself (and still act as terminated), this
made it more feasible to use these singular arrays as a way of
generically getting GC behavior for small REBVAL-sized items.  This
was used in places like function bodies, or to make a form of HANDLE!
that could participate in GC.

Another place where lifetime management for individual REBVAL-sized
things would be helpful is the external C API (or "RL_API").  Before it
spoke in terms of proxy REBVALs called "RXIARG"s which had no design
for their lifetime.  If the REBVALs used in the API lived inside of
something like a singular REBSER, they could participate in GC, and
have their lifetime attached to something like the lifetime of the
function frame that requested them.

Although a singular array could be used for this purpose, the singular
array has a property that's not really needed...which is to have
termination via an END marker.  The REBVAL is isolated and does not
have to appear as a BLOCK! of values.  This raised the possibility of
coming up with another layout for the bits inside the REBSER...which
could actually lay two entire REBVALs inside of a REBSER node.

This new structure (known as a "pairing") meant the API exposed values
could also have a "meta value".  This value could be anything from an
INTEGER! reference count, to a HANDLE!, to a FRAME! of the function
which--when it is done running--should indicate the garbage collection
of the value and its pointer.  But also, it gave a reusable "exactly
two values long" form of series that could be used with very little
overhead--and which would be unlikely to fragment memory, because the
size was the commonly-used size of a REBSER.

This commit goes ahead and makes the FRAME!-based GC of these values
work (it was just pseudocode before).