Get rid of RL_API doc generator, more tab excising
The comments in %a-lib.c were structurally analyzed to produce a text
file for documentation of the RL_API that nobody has probably ever
actually read (at least not in many years).  It used an ad-hoc parsing
technique when there is now a much better C-language parser from
@CodeByBrett.

(Also, as basically none of the RL_API routines are being kept as they
were, the arbitrary restrictions on the comment blocks for the format
was really just in the way.)

If documentation is desired, it should be done as an extension of the
current source analysis parser.  This will be more appropriate when
the RL_API has been stabilized.

This also adds to %common.r an override of `tab` which causes an error,
to try and prevent putting tab characters in the files that the tools
generate.  (They can still use tab escapes in strings or other means,
but this cuts down on the cases...hopefully eliminated eventually.)

Also macro++ changed to PROCEDURE for correct bootstrap.