Start minifying the codec code and studying dependencies
The codec code was a fairly skeletal concept of linking C code up to
decode a BINARY! as strings or images, or vice-versa.  The C code was
supposed to be only minimally dependent on Rebol, so as not to need
too much intimate knowledge of the datatypes.  This separation was
somewhat illusory, as all that really was happening was that the guts
of those datatypes were mirrored into a REBCDI structure in a very
convoluted way.

This commit moves things toward a simpler perspective, where a codec
is basically just a FUNCTION! which transforms things to or from a
BINARY! value.  Because this would require a codec to at minimum be
able to use the RL_API to examine and construct values, this does some
reorganization and documentation of that dependency on %u-jpg.c, which
seemed to be made with a particular sensitivity to not pulling in the
full dependency of %sys-core.h (while other files do).