Simplify REBIXO so that it can be on by default in the C++ build
As part of the general problem of various points in the code "missing"
thrown values, or accidentally passing THROWN_FLAG or END_FLAG to
routines that were expecting ordinary integers, a class called
Reb_Indexor was introduced.

This was before include files had been broken out as fine-grained as
they are today, and also when the idea of a C++ class appearing in
one of the ordinary headers was "scary".  So it was broken out into
a separate file called %sys-do-cpp.h, which was supposed to be a
generic set of C++ enhancements to %sys-do.h

The class was a bit overkill with trying to produce information in the
debugger, and having a binary footprint that was different from that
of the C build.  Because a C-built Rebol core couldn't be used with
a C++-built client compatibly, the checks were disabled after they
had been used to find problems.

This revives the checks and gives indexors their own file, in
%sys-indexor.h - adding further safety by making indexors compile-time
incompatible with NOT_FOUND and UNKNOWN.  This required adapting the
PARSE code to use the pairing of END_FLAG with THROWN_FLAG, instead
of NOT_FOUND and THROWN_FLAG.

The check could be made even stronger if it ensured that no indexor
was coerced into an integer value unless it had been first compared
against both THROWN_FLAG and END_FLAG.  Although that would be more
complex, it would prevent certain categories of bugs.