Reset P_OUT slot to END if not THROWN in parse
The internal logic of the R3-Alpha PARSE was exposed somewhat by
recasting it as recursive calls to a variadic native called SUBPARSE.

Nested calls to SUBPARSE from inside PARSE reuse the same output cell,
and in the case of thrown values this means that there is no need
to copy the value up the stack.  However, if a value is not thrown and
SUBPARSE's result is merely being converted into a temporary piece of
status information, then the cell may be reused in a later call to
SUBPARSE.

To try and keep things more tightly under control (and take advantage
of the fact that D_OUT is an END cell initially for any native call),
there is an assertion in subparse that the incoming output cell is
an END.  This resets the cell after a subparse once the necessary
information has been extracted (and when it's realized it's not a
value that needs to be preserved and thrown).