Simplify EMIT-LINE in %make-boot.r (among other simplifications)
The %tools/ code for bootstrap is written in Rebol, and produces C.
It's old and byzantine and a pain to deal with any time it is being
extended.  In particular, EMIT-LINE was a ridiculous abstraction over
emitting a single line of text, much more confusing than just writing
the emit at the callsite.  Here's the original R3-Alpha code, reprinted
for posterity:

    emit-line: func [
        prefix word cmt /var /define /code /decl /up1 /local str
    ][
        str: to-c-name word
        if word = 0 [prefix: ""]
        if not any [code decl] [
            either var [uppercase/part str 1] [uppercase str]
        ]
        if up1 [uppercase/part str 1]
        str: any [
            if define [rejoin [prefix str]]
            if code   [rejoin ["    " prefix str cmt]]
            if decl   [rejoin [prefix str cmt]]
            rejoin ["    " prefix str ","]
        ]
        if any [code decl] [cmt: none]
        if cmt [
            len: 31 - length? str
            loop to-integer len / 4 [append str tab]
            any [
                if define [repend str cmt]
                if cmt [repend str ["// " cmt]]
            ]
        ]
        append str newline
        append out str
    ]

Note that to get "#define PRE_X 10" (for instance) you have to write:

    emit-line/define {#define PRE} "X" 10

Which is including the #define as part of the "prefix" (as if it's the
prefix of the name, composed with X) and then 10 is the "comment".
Interpreting other callsites is equally opaque, such as:

    emit-line "RE_" reform [err "=" code] reform [code mold val]

Though what the code is ultimately doing is still a bit pathological,
this commit at least cleans up the unnecessary confusion of EMIT.