No function arg enums for natives, factor out text emitter
Because native definitions were in a separate file in R3-Alpha (and
not blended into the C source in comments) it was difficult to keep
track of numeric indexes, especially if a routine had a large number
of parameters and refinements.  A somewhat clunky way of dealing with
this was to abstract the argument IDs into enums, and the refinements
into a bit mask.

Ren-C has a better answer with the PARAM() and ARG() macros, which can
be optimized out to simple integer references into the frame.  At the
moment, however, this can be somewhat redundant with actions by
repeating the paramlist in multiple dispatchers' switch() statements.
But for natives, there's no reason to use this approach...so the two
functions that did (CHECKSUM and REQUEST-FILE) were changed to use
the new method.

Also factors out the EMIT-XXX routines so that both %make-boot.r and
%make-headers.r use common names and code for the functions.  Right
now it's a little dicey because %make-headers.r does some #ifdef
merging by looking at the emit buffer after the fact, and has to
abstract that across two different buffers at the same time...but
it's a move toward improvement.