Auto-generate of PARAM() and REFINE() defs for natives/actions
This commit began as an attempt to simplify the %tmp-funcargs.h code
and eliminate Find_Refines().  It evolved into a general solution for
auto-generating the parameter and refinement lists for natives and
actions.  So instead of writing:

    //
    //  foo: native [arg /ref ref-arg]
    //
    REBNATIVE(foo) {
        PARAM(1, arg);
        REFINE(2, ref);
        PARAM(3, ref_arg);
        // blah blah ARG(arg), REF(ref), ARG(ref_arg);

A macro is auto-generated for you with the correct indices and names.

     //
     //  foo: native [arg /ref ref-arg]
     //
     REBNATIVE(foo) {
         INCLUDE_PARAMS_OF_FOO;
         // blah blah ARG(arg), REF(ref), ARG(ref_arg);

The style of macro taking no arguments was chosen to help call it out
as unique and having no runtime behavior (in the release build).  Also
if it were folded into the REBNATIVE(...) macro, then it would have to
include the `{` character in the macro, which is kind of junky and
alien to C like doing `#define BEGIN {`, `#define END }`.  :-/

But more importantly, this means the macros can be used anywhere and
multiple times, meaning that now actions can include their parameter
lists in the several different switch() statements they appear.
Whether that's how actions will be done in the future is not really
the point, but rather that there's a much less brittle connection
between the parameter order and naming than before.

Because this supersedes the Find_Refines() method completely, it meant
the cases where the argument bit flags were used as flags to pass to
routines outside of the native or action had to be adapted.  This
pared down those flags to a minimal temporary set, but leaving it for
future review.

The modifications to the build process and simplifications are gearing
up for a large change to the "life support" of kludgey C code that
currently lets the evaluator actually do things.  So since this touched
a lot of lines, things like commented-out code were just deleted in
the process.

While all D_REF() instances were eliminated, this did not attack the
D_ARG() cases inside of action switch statements--which are fairly
entrenched because they share handling of arguments which do not
have common names in the action dispatchers.  However, all natives
were converted to the new convention.