Reverse the sense of the EVALUATED value bit to UNEVALUATED
Noticing a pattern in problems, Ren-C added the feature that it would
be possible for a function FOO with an evaluated argument to tell the
difference between being called as `FOO [y]` vs. `FOO SECOND [[x] [y]]`
or `BLK: [y] | FOO BLK`.  The reason this distinction was important
was to give errors like:

    >> if [x] [print "oops"]
    ** Script error: Literal block used as conditional [x]
    ** Where: if
    ** Near: if [x] [print "oops"] ??

The initial experiment to add this feature added a bit to REBVALs
called VALUE_FLAG_EVALUATED.  Every function evaluation or variable
fetch would set this bit, while literal values or QUOTE would clear
it off.  It was a property of the functions QUOTE and SEMIQUOTE,
which had to be tested for explicitly on each evaluation.

(Note: This opens a bit of a can of worms, since the QUOTE native
is no longer purely compatible with a user function that implements
QUOTE by just returning a GET-WORD!-styled argument.  That could
be addressed by allowing user functions to mark their return values
as literal, either with a RETURN/LITERAL or just saying that a function
will always return what is effectively a literal or not.)

In any case, the feature was added before the R_XXX function overhaul,
which makes good use of the ability of functions to do one level of
optimized switch statement after dispatch.  Hence there's no need
for each and every evaluator step to examine the function pointer and
do two comparisons against QUOTE and SEMIQUOTE to decide what to do
about the bit...instead, those natives can just return a different
value as R_OUT_UNEVALUATED.

In addition to that optimization, this changes it so that the default
for all the initialization operations is to make evaluated values.
So the flag in the positive sense is VALUE_FLAG_UNEVALUATED instead of
VALUE_FLAG_EVALUATED.  Technically speaking this doesn't matter that
much since the optimizer would fold together operations like
`SET_BLANK(val); SET_VAL_FLAG(val, VALUE_FLAG_EVALUATED);` into a
single assignment of the header bits, but given that the unevaluated
bit is the "special" case there are fewer mentions by reversing the
sense of the bit.