Unpoison s->do_count before reassigning
Conceptually, `s' is free'd and shouldn't be accessed. But the
in the current implementation, freeing a node is just returning
it to the memory pool for reuse, thus it's still under the
control of the application. However, when coupled with memory
poisonning, this part of the memory will be poisoned until it's
reallocated for other usages.

But this is only for debugging purpose, just unpoison and repoison
it again.