Eliminate all eol-wsp errors from the C files
Although there was an end of line whitespace checker in the source
analyzer, it was not applying the test to comments or empty lines.
Notably, the emitted comment headers for natives had comment lines
which had spaces on the end.

Using the information from the enhanced check in the commit by
@CodeByBrett, this gets rid of the whitespace reported.