Tweak voids in REMOVE-EACH/"UNTIL", use data stack in MAP-EACH
In Ren-C, CONTINUE has a /WITH refinement.  The meaning of calling
CONTINUE/WITH a value is to do exactly the same thing as if the loop
body had finished normally and evaluated to that value:

    >> map-each x [1 2 3] [if x = 2 [continue/with 666] x + 100]
    == [101 666 103]

(Note: BREAK also has a /WITH refinement--instead of /RETURN--with a
similar semantic of acting like the body returned the value...it just
also stops the loop after that.)

The next rule is that a CONTINUE *without* a /WITH refinement acts
exactly like if the loop body had no result.  Hence if one desires
the following:

    >> foo: [1 2 3]
    >> remove-each x foo [either x = 2 [true] [continue]]
    >> probe foo
    [1 3]

...that implies that if the loop body of REMOVE-EACH evaluates to no
value, it is "opting out" of the removal.  Hence it will act the same
as a FALSE or BLANK! value would.

    >> foo: [1 2 3]
    >> remove-each x foo [print "hi"]
    hi
    hi
    hi
    >> probe foo
    [1 2 3]

This fixes what was an safety asserting case (logic testing on a void)
by resolving it to be consistent with this invariant.  It also goes
ahead and updates the behavior for LOOP-WHILE and LOOP-UNTIL (aka
"UNTIL") to similarly allow voids to opt out, to give the same
consistency with CONTINUE.  Hence LOOP-UNTIL [print "hi"] is an
infinite loop, as LOOP-UNTIL [false] was, and LOOP-WHILE [print "hi"]
is an infinite loop, as LOOP-WHILE [true] is.

(Note: The somewhat unusual impliciation is that CONTINUE/WITH TRUE
will break a LOOP-UNTIL, and CONTINUE/WITH FALSE will break a
LOOP-WHILE, but it makes sense as their body and condition are the
same thing.)

In addition to resolving this issue, the commit relaxes the requirement
that you cannot guard a series from garbage collection unless it is
managed by the garbage collector.  The way it is relaxed is that if
the series is an array--and unmanaged--then there is a special form
of guarding of its *contents*.  This was initially used to enhance the
handling of MAP-EACH's temporary buffer, before realizing that MAP-EACH
really should be using the data stack.  The data stack already offers
the necessary protection, as well as helping to generate an array of
the right size.  This is why REDUCE and other routines use it.