Make REBRIN an array instead of a custom C struct
The FFI implementation originally used dedicated C structures, which
were allocated from a specialized memory pool.  This required some
fairly invasive changes to the GC in order to put these structures
into the graph of "live" references for the system.  Because these
specialized structures had their own GC code, it was also not as
complete as the system's ordinary code...meaning deeply nested
structures could overflow the stack.

Rather than keep expanding or bulletproofing that approach, this moves
in the other direction to get rid of the specialized code.  This
commit contains a first step--which is to turn the REBRIN C struct
into an ordinary Rebol array of 8 values.  It makes use of the new
GC-able HANDLE! type, which allows the FFI-allocated structures to be
cleaned up.

(It means a very slight amount more storage needed per-routine, at
the savings of code complexity and better robustness.  But Given the
relatively small number of routines made in most sessions, the extra
bytes are likely more than compensated for by the fewer bytes of code!)