Make Reb_Field an array instead of a C struct
Following on the change of the REBRIN to an array of ordinary Rebol
values, this changes the `struct Reb_Field` into an alias for an
ordinary array, REBFLD.  This change allows all the FFI-oriented code
to be removed from %m-gc.c...simplifying the garbage collector, and
meaning that non-FFI builds aren't paying for code in the core related
to that.  It also focuses the core on enhancing generalized features
that provide benefit across the board, without the risk of brittle
specialized code for the FFI structures.

(Just as an example, the struct recursive GC marking could not use the
same queue as the REBSER mark queue.  Hence, a deeply nested struct
could in theory overflow the stack.  The work done on making this not
a problem for REBSER-based GC could not be reused.)

One feature that could not be preserved easily in this change was the
ability to tunnel REBVALs into FFI routines.  Although your average
FFI routine pretty much does not--by definition--know what to do with
a REBVAL (as it is "foreign", not extension code)...it could be useful
in tunneling a REBVAL back into Rebol via a callback.  But having a
REBVAL's bit pattern living in binary data would require specialized
marking code in %m-gc.c, and the point of this commit is to get rid
of that.  If this feature is important, it will need to be done some
other way.

Similarly to the change in REBRIN, this means that there are slightly
more bits used to define the schemas of STRUCT!.  However, since
schemas are shared between structs of the same type, the overhead of
this is proportional to the total number of structs definitoins in the
system...and not the struct instances.  Hence the reduction in code and
complexity almost certainly means less bytes overall.

While the change was being made, the opportunity was taken to switch
the parameter building code to use the data stack--as opposed to
pre-guessing the size of certain arrays in a way it might get wrong.
If the code were converted to Rebol, this would correspond to doing
something more like a COLLECT/KEEP.

This is deliberately a first cut that just gets it going, which will
be refined in further commits.  This should make it easier to find
where something went wrong if there is a problem.