Replace "safe trash" concept with "unreadable blank"
In order to not contaminate the enumerated type list with a separate
debug-only concept of a TRASH! value, the same REB_XXX enum value was
used to represent voids, as well as both "GC safe" and "GC unsafe"
trashes.  In the release build, a safe trash would become a void.

In the modern world of Ren-C, this is not a very safe form of trash...
because most blocks can't have voids in them.  It's better for the
release build to choose a BLANK!.  To help with the language a bit,
the "safe trash" is considered instead an "unreadable blank"...it
acts as a blank does for the GC purposes, but cannot have its VAL_TYPE
examined.  It can only be tested for with IS_BLANK_RAW().

So overall this is clearer, because from a release build's point of
view, it's a blank...so taking the "trash" name out and thinking of
it as a modified blank is better.  However, for performance reasons,
it still uses the same REB_XXX value and flag as trash does in the
debug build, which keeps the debug build from getting any slower.

This pushes the decision through, and tries to tidy up some of the
usages.  Reviewing that raises an uncomfortable observation about
using Do_Core() to do evaluations into any user exposed value slot...
including argument slots in frames.  This will be addressed in a later
commit.