Assert if native attempts to evaluate directly in FRAME!
It is already prohibited by a heavy debug check (which is not enabled
by default) that you cannot use a cell inside an array as a destination
for an evaluation.  This is because the act of evaluating arbitrary
code itself may resize and reallocate an arbitrary array.  This could
move the intended destination pointer, but the evaluation would still
be wanting to write into the stale cell.

This has a side benefit that no user-exposed value in a BLOCK! or
OBJECT! can be seen in an intermediate state of construction.  In fact,
the first thing an evaluation into a cell does is to write an END
marker in that cell, so as to ensure the cell is GC-safe (and since
it's not a valid value to output, this means it's possible to detect
if the native never actually wrote a result.)  But this should never
be done into a BLOCK! or OBJECT!; it will always go through an
intermediate cell, and then move the value into it.

But there was a case that wasn't being checked, and in fact was used
as an optimization which was actually incorrect.  This is when
running native functions would DO directly into one of their arguments.
While this is fine for the evaluator to do when it is fulfilling
arguments of a not-yet-user-exposed frame...once the function
has started running its arguments are exposed via WORD!s (and now
via FRAME!s).  So it could potentially be the case that the arbitrary
code running in the DO gets its hands on the pointer to that cell
while the code is running.

Besides adding a check for this, it also adds a test for the use of
the evaluation cell in routines that are using the "one argument"
optimization.  This is how a native that takes only one argument
(not including the RETURN, which is omitted in release builds) will
use the evaluation cell to hold its single arg.